✨ volatile write
with the previous approach, it was possible for the compiler to
'optimise' and remove the write functionality, as to write the text was
just being saved to the buffer location, and not accessed from there.
this may have created the illusion that it wasn't being used; hence good
to remove to optimise.

volatile crate helps prevent that.